+++ packages/resolvers/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@vue-hooks-plus/resolvers",
-  "version": "1.2.1",
+  "version": "1.2.3",
   "description": "",
   "type": "module",
   "main": "dist/index.cjs",
