+++ packages/use-worker/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@vue-hooks-plus/use-worker",
-  "version": "1.0.1",
+  "version": "1.0.3",
   "description": "",
   "files": [
     "dist",
